requirements.txt
requests~=2.32.3
gradio~=4.41.0
loguru~=0.7.2
python-daemon
langchain~=0.2.14
langchain-core~=0.2.33
langchain-community
openai~=1.42.0
config~=0.5.1
beautifulsoup4~=4.12.3
django-scheduler~=0.10.1